+++ src/onegov.org/onegov/org/tests/test_views.py
@@ -13,7 +13,7 @@ from libres.modules.errors import AffectedReservationError
 from lxml.html import document_fromstring
 from onegov.core.utils import Bunch
 from onegov.form import FormCollection, FormSubmission
-from onegov.libres import ResourceCollection
+from onegov.reservation import ResourceCollection
 from onegov.newsletter import RecipientCollection
 from onegov.org.testing import Client
 from onegov.org.testing import decode_map_value