+++ src/onegov.org/onegov/org/views/allocation.py
@@ -3,7 +3,7 @@ import morepath
 from libres.db.models import Allocation
 from libres.modules.errors import LibresError
 from onegov.core.security import Public, Private
-from onegov.libres import Resource, ResourceCollection
+from onegov.reservation import Resource, ResourceCollection
 from onegov.org import OrgApp, utils, _
 from onegov.org.elements import Link
 from onegov.org.forms import (