+++ src/onegov.org/onegov/org/views/reservation.py
@@ -8,7 +8,7 @@ from libres.db.models import Allocation, Reservation
 from libres.modules.errors import LibresError
 from onegov.core.security import Public, Private
 from onegov.form import FormCollection, merge_forms
-from onegov.libres import Resource
+from onegov.reservation import Resource
 from onegov.org import utils
 from onegov.org.elements import Link
 from onegov.org.forms import ReservationForm